Refactor and add field-centric control
The IMU was initialized and the parameters were added.
The DriveMotor class was redone such that it has a static method that
creates a new DriveMotorEx object rather than using a constructor that
creates a parameter; this is cleaner and returns an object rather than a
parameter.
The DriveMotor class was also moved to the TeleOp.utils package, for
organization purposes.
TODO: add comments explaining how the rotX and rotY variables are
calculated.