Add telemetry and fix small things
This moves the comments on lines 35 and 36 all to line 34, for
cleanliness.
It also changes the imu parameters to be variables instead, to make it
easier to change them.
Telemetry was added to the moveRobot method to display the power of each
motor and the heading of the IMU, for testing purposes.
DcMotorEx was changed to DcMotor, as I am not sure if our drive motors
are compatible with DcMotorEx. This will have to be tested on the robot.